refactor: move previous-environment helpers into the new controller
Relocate getNoteDrySha, getEffectiveHydratedDrySha, checkCommitStatusesPassing,
and isPreviousEnvironmentPending into the PreviousEnvironmentCommitStatus
controller. Same package, so call sites are unchanged — behavior-preserving.
